Extract MapCoordsRegion from CellRegion.
